// src/types.ts
export type FlexDirection = 'row' | 'row-reverse' | 'column' | 'column-reverse';
export type FlexAlign =
  'auto'
  | 'baseline'
  | 'center'
  | 'flexStart'
  | 'flexEnd'
  | 'spaceAround'
  | 'spaceBetween'
  | 'stretch';
export type FlexJustify = 'center' | 'flexStart' | 'flexEnd' | 'spaceAround' | 'spaceBetween' | 'spaceEvenly';
export type FlexWrap = 'noWrap' | 'wrap' | 'wrapReverse';